Optimize single entries in `series` and `parallel`.
When there's only one entry to either the function `series` or `parallel` no action need be taken since only the original middleware should be run. This is a bit of a microoptimization that pevents a couple of unnecessary function calls. A case for the 0-length middleware list could also be an option.